feat(theme): migrate theme from V2 site.
The theming from the Astro V2 site has been imported into this website
to allow me to clone the style of the previous site. This also includes
adding the Inter font from Google Fonts. A self-hosted option may be
explored in the future to remove the dependency on Google, which some
may find undesireable.